==> 02_compas/02_compas_scripts/01_compas_load.Rmd <==
---
title: "Compas - Load Data"
author: "Simon Schrauth"
date: "`r Sys.Date()`"
output: html_document
---
# Load Compas Dataset

## Goal

After executing this script, the compas dataset should be loaded.

## Install packages

```{r}
pacman::p_load(RSQLite,
               tidyverse, 
               here)
```

## Database Shizzle

```{r}
# con = dbConnect(SQLite(), "../01_compas_data/01_compas_data_raw/compas_data.db")
# 
# dbListTables(con)
# 
# dbGetQuery(con, "SELECT * FROM prisonhistory")

```

## Load data from csv-file

```{r}
compas_data_recid_raw = read_csv(file = here("02_compas", "01_compas_data", "01_compas_data_raw",
                                             "compas-scores-two-years.csv"),
                         col_names = TRUE,
                         na = c(""))

compas_data_recid_raw
```

```{r}
compas_data_voil_raw = read_csv(file = here("02_compas", "01_compas_data", "01_compas_data_raw",
                                            "compas-scores-two-years-violent.csv"),
                         col_names = TRUE,
                         na = c(""))

compas_data_voil_raw
```


## Save data

```{r}
saveRDS(compas_data_raw, 
        file = here("02_compas", "01_compas_data", "02_compas_data_processed",
                    "01_compas_data_recid_load.rds")
        )

saveRDS(compas_data_raw, 
        file = here("02_compas", "01_compas_data", "02_compas_data_processed",
                    "02_compas_data_viol_load.rds")
        )
```
