==> 02_compas/02_compas_scripts/03_compas_model_preperation.Rmd <==
---
title: "Compas - Model Preparation"
author: "Simon Schrauth"
date: "`r Sys.Date()`"
output: html_document
---
# Preparation of the Compas Dataset for Modeling

## Goal

After executing this script, the compas dataset should be ready to be modeled as deep neural network in keras.

## Preparations

### Load R packages

```{r}
pacman::p_load(tidyverse, 
               here,
               reticulate)
```

### Activate Python environment

```{r}
reticulate::use_condaenv(condaenv = "Masterarbeit", 
                         required = TRUE
                         )
```

### Import Python packages

```{python}
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.compose import make_column_transformer
from sklearn.preprocessing import StandardScaler, OneHotEncoder
from sklearn.compose import make_column_selector
from sklearn.model_selection import train_test_split
from sklearn import preprocessing
```


### Load data

```{r}
compas_data_clean = readRDS(file = here("02_compas", "01_compas_data", "02_compas_data_processed",
                                      "03_compas_data_clean.rds"))
```


### Transform to python data frame

```{python}
compas_data_clean_pre = r.compas_data_clean
```

### Function for Standardization of the numeric features & One-Hot-Encoding of the categorical features

```{python}
# function for cleaning column names of add-ons from transformation steps (help function for feature_transformator below)
def clean_colnames(colnames):
  
  # delete/replace some specific strings
  clean_names = list(map(lambda x: x.replace("remainder__", "").replace("onehotencoder__", "").replace("standardscaler__", "").replace("-", "_").replace(" ", "_").replace("[", "").replace(")", ""), colnames))
  
  return(clean_names)



# actual transformation function
def compas_feature_transformator(df):
  
  # define transformation for each data type
  column_trans = make_column_transformer(
    
      (OneHotEncoder(), make_column_selector(dtype_exclude = float)),      # dummy-encoding of categorical features
      
      (StandardScaler(), make_column_selector(pattern = "^(?!two_year_recid$).*$",dtype_include = float)),     # standardization of numerical features except target feature
      
      remainder = 'passthrough'   # target feature remains in its original binary state
    )
  
  # create new dataframe with transformed features and column names 
  new_df = pd.DataFrame(column_trans.fit_transform(df),
                        index = df.index,
                        columns = clean_colnames(column_trans.get_feature_names_out())
                        )
  return(new_df)

```

### Feature Transformation 

```{python}
compas_data_clean = compas_feature_transformator(compas_data_clean_pre)
```


## Preparation of the dataset

### Ratios for training, testing and validation set
```{python}
compas_train_ratio = 0.75
compas_validation_ratio = 0.15
compas_test_ratio = 0.1
```


### Train-test-split for training and testing datasets
```{python}
compas_X_train_raw, compas_X_test_raw, compas_y_train_raw, compas_y_test_raw = train_test_split(compas_data_clean.iloc[:,:-1], 
                                                                                                compas_data_clean.iloc[:,-1], 
                                                                                                test_size=compas_test_ratio,
                                                                                                random_state=42,
                                                                                                shuffle = True)
```

### Train-test-split for training and validation datasets
```{python}
compas_X_train_raw, compas_X_val_raw, compas_y_train_raw, compas_y_val_raw = train_test_split(compas_X_train_raw, 
                                                                                              compas_y_train_raw, 
                                                                                              test_size=compas_validation_ratio/(1-compas_test_ratio),
                                                                                              random_state=24,
                                                                                              shuffle = True)
```

### Reset all indices
```{python}
compas_X_train = compas_X_train_raw.reset_index(drop=True)
compas_y_train = compas_y_train_raw.reset_index(drop=True)

compas_X_val = compas_X_val_raw.reset_index(drop=True)
compas_y_val = compas_y_val_raw.reset_index(drop=True)

compas_X_test = compas_X_test_raw.reset_index(drop=True)
compas_y_test = compas_y_test_raw.reset_index(drop=True)
```

### Save column names for interpretation
```{python}
compas_col_names_enc = compas_X_train.columns.values.tolist()
```

```{python}
compas_X_train
compas_y_train

compas_X_val
compas_y_val

compas_X_test
compas_y_test

compas_col_names_enc
```

### Save as RDS
```{r}
compas_X_train = py$compas_X_train
saveRDS(compas_X_train, 
        file = here("02_compas", "01_compas_data", "02_compas_data_processed",
                    "04_compas_data_model_X_train.rds")
        )

compas_y_train = py$compas_y_train
saveRDS(compas_y_train, 
        file = here("02_compas", "01_compas_data", "02_compas_data_processed",
                    "05_compas_data_model_y_train.rds")
        )

compas_X_val = py$compas_X_val
saveRDS(compas_X_val, 
        file = here("02_compas", "01_compas_data", "02_compas_data_processed",
                    "06_compas_data_model_X_val.rds")
        )

compas_y_val = py$compas_y_val
saveRDS(compas_y_val, 
        file = here("02_compas", "01_compas_data", "02_compas_data_processed",
                    "07_compas_data_model_y_val.rds")
        )

compas_X_test = py$compas_X_test
saveRDS(compas_X_test, 
        file = here("02_compas", "01_compas_data", "02_compas_data_processed",
                    "08_compas_data_model_X_test.rds")
        )

compas_y_test = py$compas_y_test
saveRDS(compas_y_test, 
        file = here("02_compas", "01_compas_data", "02_compas_data_processed",
                    "09_compas_data_model_y_test.rds")
        )

compas_col_names_enc = py$compas_col_names_enc
saveRDS(compas_col_names_enc, 
        file = here("02_compas", "01_compas_data", "02_compas_data_processed",
                    "10_compas_data_model_col_names_enc.rds")
        )
```
